Fix for growing Offer/Answers
In webrtc's unified-plan semantics, instead of adding tracks to streams, you
operate mostly in transceivers. When hanging up a stream, the transceiver
should be Stopped and the track removed with the RemoveTrackNew() method, which
should be used over the plan-b RemoveTrack() method.

Testing locally confirms that after hanging up a stream, the transceivers are
now being stopped, and no longer included in Offer/Answers. Additionally,
previously stopped Transceivers can be used for new streams.

With this change, we no longer need to force a PeerConnection reset, and it should
also improve stream latencies.